Добавлено: - Отображение системных уведомлений - Отображение уведомлений об оплате
- В работе -> пагинация

+++ app/src/main/java/kz/smartideagroup/pillikan/content/home/HomeFoundationRepository.kt
@@ -9,7 +9,7 @@ class HomeFoundationRepository(application: Application): BaseRepository(applica
         return userSession.getUserName().toString()
     }
 
-    fun getUserBalance(): Int {
+    fun getUserBalance(): Float {
         return userSession.getBalance()
     }
